// src/app/models/card-params.ts
export interface CardParams {
    num?: Number;
    offset?: Number;
    name?: String;
    fname?: String;
    id?: Number;
    konami_id?: Number;
    type?: String;
    atk?: Number;
    def?: Number;
    level?: Number;
    race?: String;
    attribute?: String;
    link?: String;
    linkmarker?: String;
    scale?: Number;
    cardset?: String;
    archetype?: String;
    banlist?: String;
    sort?: String;
    format?: String;
    misc?: String;
    staple?: String;
    has_effect?: Boolean;
    startdat?: String;
}